feat: Add word count functionality
Now it's possible to count the number of words in an English text using NLTK[1].
This functionality utilizes NLTK's ability to classify words and then count the
relevant words.

Additionally, two corpora have been added to the project: Punktz and
averaged_perceptron_tagger. These corpora are used for word classification.

[1] https://www.nltk.org/